\name{testMACDrules-package}
\alias{testMACDrules-package}
\alias{testMACDrules}
\docType{package}
\title{
MACD examples
}
\description{
Short examples for MACD, testing SMA and EMA options
}
\details{
\tabular{ll}{
Package: \tab testMACDrules\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-09\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Diego J Bodas Sagi

Maintainer: Diego J Bodas Sagi <diegobodas@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ MACD }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
##see papers
}
